GUACAMOLE-168: Don't require virtual console
When running X.org in a container the service will fail to start
properly and report that `/dev/tty0` is not present or could not be
acquired.

It's possible to work around this issue by enabling console pass-through
from the host system -- in Docker using the `--privileged` flag. However
this solution is not ideal as it requires granting pretty open access to
the host system. Additionally, the x-server driver does not actually
need access to the virtual console.

When X.org starts up it probes all devices to check what hardware is
necessary for each device. Unless the devices specifies, X.org assumes
that a virtual console is necessary. This behavior can be changed by
registering a `driverFunc` which responsds to the
`GET_REQUIRED_HW_INTERFACES` operation with `HW_SKIP_CONSOLE`.

Some discussion about this behavior. The "dummy" driver for x-server was
updated to also remove the need for a virtual console, which caused an
issue for users using screen readers, as when the dummy driver was the
_only_ driver then no keyboard input was registering. This issue is not
relevant to the Guacamole x-server driver but this has useful
information and background for this.
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=696965

This change is modelled after the way the dummy driver behaves to turn
off the need for a virtual console.
- https://github.com/Xpra-org/xf86-video-dummy/blob/master/src/dummy_driver.c

With this change there is no longer a need for exposing the host's
virtual console to the container which runs X.org with only the
Guacamole driver.