GUACAMOLE-168: Detect and use libXfont2 if available
The libXfont library does not seem to be available on some systems
unless using version 2.

The current use of libXfont is minimal. This change updates the
configure script to check for the libXfont2 library and link against
that instead along with defining a new `HAVE_XFONT2` to be used within
the driver code.

This change was inspired by:
https://cgit.freedesktop.org/xorg/driver/xf86-video-qxl/commit/?id=e13d28ee5d8724fc4b22f26bce01a7d36355f272